refactor: Separate validator module into distinct domains
