+++ src/flycatcher/validators/__init__.py
@@ -1,0 +1,14 @@
+"""`col()` alias and validator DSL public API for the flycatcher package."""
+
+from .core import BinaryOp, FieldRef, UnaryOp, ValidatorResult, col
+from .string import StringAccessor, StringOp
+
+__all__ = [
+    "FieldRef",
+    "BinaryOp",
+    "UnaryOp",
+    "col",
+    "ValidatorResult",
+    "StringAccessor",
+    "StringOp",
+]
